дописал логику PaymentService

+++ InventoryService/Dto/CreateGoodDto.cs
@@ -1,3 +1,3 @@
 namespace InventoryService.Dto;
 
-public record CreateGoodDto(string Sku, string Name, int ItemPrice, int? Quantity);
+public record CreateGoodDto(string Sku, string Name, decimal ItemPrice, int? Quantity);
