+++ InventoryService/Dto/GetGoodDto.cs
@@ -1,3 +1,3 @@
 namespace InventoryService.Dto;
 
-public record GetGoodDto(string Sku, string Name, int Quantity, int ItemPrice);
+public record GetGoodDto(string Sku, string Name, int Quantity, decimal ItemPrice);
